Add support for pulling dumps in the FR policy
The transport.sh script can do both push and pull, but we need to
tell it what to do and, in case of pull, which hosts to pull the
dumps from. This is determined from the contents of the FR
config.

Also, instead of the superhub trusting (the keys of) the feeders,
it's the feeders that need to trust (the key of) the superhub.

Ticket: ENT-4499
Changelog: None
(cherry picked from commit f5bc8645b2b96c6d514b564de653d04005bacbce)