Setup our own symlink for Python interpreter and use it
Different operating systems or distributions have different
versions of Python. And while our Python scripts can run with
various versions of the interpreter, we need to know the exact
path to it. Let's just use a common path '@bindir@/python' and
make sure it is a valid symlink to a Python interpreter we found
in the system (if any).

Ticket: ENT-4682
Ticket: ENT-4668
Changelog: Title