Fixup Python for package module tests using it
Package modules using Python now use the symlink we create in
policy. In the tests, there's no agent run to set this symlink up
so we need to work this around. For unit tests, we can execute
the package modules using system Python. For unsafe acceptance
tests, we just set the symlink in the test's init bundle.

Ticket: ENT-4682
Ticket: ENT-4668
Changelog: None