Added continual checking for policy_server state
Previously, we relied only on the presence of the am_policy_hub state file for
classification. This change regularly checks to see if the executing host is
bootstrapped to an ip found on itself, and if so it ensures the state file is
present. This helps to avoid accidental declassification from the accidental
removal of the state file. It uses the same logic that is used during bootstrap.

Ticket: CFE-3073
Changelog: Title
(cherry picked from commit 489c55cc892eab18f98f986c45bb893c85dda92e)