Fixed problem with postgresql.conf not getting configured
The postgresql.conf file was not being edited as expected, when working with
namespaces you must be careful to remember and include the namespace when
passing names of variables. This change added the namespace to the
set_line_based edit_line bundle so that the variable could be properly resolved
to the expected values.

Ticket: ENT-4744
Changelog: None
(cherry picked from commit 16027eb620418b2905beeaf5be473b5238e569a4)