Do not require specific SSH key type for FR
Systems may have various restrictions for SSH key types they
use. Let's just leave it to the system configuration to
decide. One of the problematic cases here is RHEL 6 which has an
older version of SSH that doesn't support the specific key type
we were trying to use.

Because we no longer specify the key type, we need to specify the
key file name and then make sure it is being used (SSH only uses
the default-named keys like id_rsa, id_dsa,...).