Add failure handling in FR import
We need to revert changes if things fail. If preparing/seting up
schema for the import fails, we abort the import process because
that means the DB is in an unexpected shape and we want to
prevent (further) damage.

If an import of data from a specific feeder fails, we just throw
the data away and revert the related schema changes. The same
applies to attaching the imported data from a specific feeder.

Ticket: ENT-4742
Changelog: none
(cherry picked from commit dc4f86030a59a66427b565556972d5e3a5b8d641)