FR, reduced default logging level for import, provided debug option
By default psql is called with --quiet and client_min_messages set to warning.
If cf-agent is called with --define DEBUG or DEBUG_config then psql is called
with --echo-all and client_min_messages set to notice.

Also, the log files are piped to gzip to reduce file size.

Changelog: none