Dump data on feeders every 20 minutes not in every agent run
Together with the rest of the ETL process this should result in
data being refreshed on the superhub every ~30 minutes.

Also note that we don't have to change the interval on the
superhub because there we want to always try to get fresh dumps
and import them. Of course, normal promise locking rules apply so
an import will only start once the previous one is done.

Ticket: ENT-4806
Changelog: none