Make sure superhub trusts all SSH keys provided by its feeders
SSH generates host keys of multiple types in modern systems. We
need to make sure all host keys from each feeder are trusted by
the superhub because only the mutual negotiation between the
superhub and each feeder determines which type of the key is used
(and thus needs to be trusted).

Ticket: ENT-4917
Changelog: Made keys of all types from feeder hubs trusted on a superhub