Fix Python 3 incompatibility in yum package module
Process output in Python 3 is of type 'bytes' which cannot be
treated as string. It first needs to be decoded into a (unicode)
string. In Python 2 it is a string which can be decoded into a
unicode object. So in both cases it is safe to do
`output.decode("utf-8")` to get a valid string-like object.

Changelog: Fixed Python 3 incompatibility in yum package module