Added and transitioned to using master_software_updates shortcut
The self upgrade policy previously fetched from a fully qualified path. This is
problematic when the hub does not use the standard WORKDIR (/var/cfengine). This
change implements a shortcut for the directory so that only the hub is required
to know the full path and switches copy_from to use paths relative to the
shortcut. Additionally, the path can be customized via augments.

Ticket: ENT-4953
Changelog: Title