modules/packages/zypper.in: Moved zypper package module errors to the cf-agent output
Errors from the package manager were redirected to stderr.
They will now be surfaced as an error in the cf-agent output.

Cf-agent expect the module to never fails, and so, should always return a 0 error code
even in failure. If the return code is not 0, for some reasons its output is dropped
and not taken into account.

This commit force most of the exit codes to 0 to make cf-agent display the newly added
error messages. The outcomes of the promises are always done by comparing multiple
module calls, so it is most likely safe to hide the package manager exit code if it
is already checked by the module itself.

Changelog: Title
Ticket: CFE-3154
(cherry picked from commit e225a15cc9039d6fc1abdebd6870aa7a9e0ebb32)