Extended watchdog for AIX
This change extends the watchdog installed when
cfe_internal_core_watchdog_enabled is defined on AIX hosts.

Previously, the watchdog would check that cf-execd was running. If it was not
running, it would start it. Now in addition to starting cf-execd if it is not
found running other symptoms of pathology are considered.

The following additional conditions are now considered
- More than one cf-execd process running

- `cf-execd: timeout waiting for output from agent` exists in outputs
  - This is logged when exec_command as defined in body executor control has not
    communicated back to cf-execd in agent_expireafter minutes.

- cf-agent process running for more than 5 minutes

- More than 3 concurrent cf-agent processes

- `cf-check diagnose $(sys.statedir)/*.lmdb` returns non-zero (indicating
  critical integrity issue)

Additionally, the watchdog now logs to `$(sys.workdir)/watchdog.log`. Note, this
files growth is currently unmanged. Left unattended over a long period of time
could result in filling the partition.

If the sum of newly introduced conditions breaches the threshold (default 0) is
breached then artifacts detailing the system state, are collected and saved to
`$(sys.workdir)/watchdog-archives/`. After information is collected, the
cfengine services are killed, outputs and LMDB databases are purged and services
are re-started. The generated archives should be submitted to the maintainers
via the Enterprise Support Portal or the public bug tracker.

This change also introduces rudimentary support for avoiding filling the disk
with watchdog archives. If the partition the archives are written to or the
temporary area contain less than 500M of free space, watchdog archives excluding
the oldest and most recent will be purged

Ticket: ENT-4995
Changelog: Title
(cherry picked from commit d6192d50dbe656f6ff9bac9da2ad88ec18039541)