Travis: Sleep 10 more seconds after killing cf-serverd
This test failed 2/5 times in the recent master builds,
because cf-serverd didn't exit in time. Increasing the
sleep from 30 to 40 seconds, to see if this helps.

We could probably work on a system which sleeps 10
seconds at a time, and checks if the process is
still alive.

Changelog: None
Ticket: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>