Excluded Enterprise federation policy parsing on incompatible versions
The federated reporting in CFEngine Enterprise leverages features that are
unavailable in versions older than 3.12.0. This change prevents the federation
policy from being parsed and generating errors on older versions (both from
including policy containing features not present, and from specifying a bundle
that is not present when the federation policy is not parsed.

3.10.x was not targeted for support because Federated Reporting was targeted for
supported use starting in 3.15.0 as 3.10.x ended it's standard support life.

Ticket: CFE-3193
Changelog: Title