Ensured directory for custom action scripts is present
In CFEngine Enterprise alerts can be based off of observations in the
infrastructure. When these alerts change state a custom action script can be
triggered in response. This change ensures that the directory where custom
action scripts are stored is available and writeable by the webserver. If this
directory is not present and writeable by the webserver, then Mission Portal
users will not be able to upload custom action scripts.

Ticket: ENT-5070
Changelog: Title