Don't change the whitespace before '=' in INI files when uncommenting
For example, if the value 'desired value' is promised for the key
'variablename' in section 'sectionname', then without this commit, the
following file:

    [sectionname]
    # variablename = default value

would be changed to:

    [sectionname]
    variablename= desired value

With this commit, it will instead be changed to:

    [sectionname]
    variablename = desired value

Which is cosmetically superior.

The change in this commit is not as crucial as the change I made in the
parent commit, but for consistency I believe the field separator regex
passed to the "col" edit_field body should be the same for both calls.