Whitespace is now allowed at the beginning of ini key-values
Previously the {manage,set}_variable_values_ini did not match on any ini files
containing whitespace at the beginning of a line, hence

[section foo]
key=value

could be handled. However, if the file looks like

[section bar]
  key=value

then the previous regex would not match.

What I can get from the INI spec, is that whitespace is optional. This commit
adds this to the regex, with an optional whitespace match at the beginning of
the line.

Changelog: Title

Ticket: CFE-3244

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>