Added inventory for CFEngine Enterprise License information
This change adds inventory of CFEngine Enterprise license utilization based on
knowledge from cf-hub. It introduces 5 new inventory items on hosts with cf-hub.

New inventory attributes:
- CFEngine Enterprise license file
- CFEngine Enterprise license utilization
- CFEngine Enterprise license expiration date
- CFEngine Enterprise licenses allocated
- CFEngine Enterprise licenses status

Because of an issue where cf-hub --show-license returns non-zero when no license
is installed, a return code of 1 is currently interpreted as a promise kept.
After ENT-5279 is resolved, the inventory will result in a promise not kept if a
license is installed, but not valid.

Ticket: ENT-5089
Changelog: Title