Suppress useless inform output from /bin/true in ec2 inventory
The commands promise /bin/true is used to delay execution of url_get until the
bundle is explicitly called (so that it doesn't get called and cached during
pre-eval). This change suppresses the useless inform output on versions 3.15.0
and later yet maintains backwards compatibility with 3.12 and 3.10 (where the
inform output can not be suppressed).

Ticket: ENT-5233
Changelog: Title