Simplify the periodic inventory refresh
The changes in cfengine/nova#1584 introduce a better way to
refresh inventory than just doing it periodically from agent
run. However, we still need to refresh the information about the
variables shown on the Inventory page in Mission Portal.

A new PHP CLI task with a more appropriate name
'inventory_variables_refresh' was introduced in
cfengine/mission-portal@d2550582, but we want to keep using the
old task for backwards compatibility.

Ticket: ENT-4864
Changelog: Inventory refresh is no longer part of agent run on the hub