Prevented inventory of unresolved variables for diskfree and loadavg
If cf-monitord is not running, or is simply unable to resolve these variables,
then they are inventoried as is. Instead of confusing the user with some
unexpanded variable, it's better to not even inventory the unresolved value.

Ticket: ENT-5190
Changelog: Title
(cherry picked from commit eb15c94847407cfa00ba00fee595c89e5bcd9770)