Stop cf-hub for the time we import data on a superhub
Data import on a superhub manipulates schemas which requires
various locks on the PostgreSQL DB. cf-hub also acquires DB locks
while it's running and so there is a risk of deadlocks. To avoid
them, let's stop the cf-hub process while the import is running
and then start it again.

This should not be a problem because the cf-hub process on a
superhub is only collecting reports from the superhub itself.

Changelog: none