Stop doing inventory refresh as part of FR import process
We have switched to doing inventory refresh as part of the import
process because we needed to do a refresh by hostkey only for the
superhub. But this is now happening implicitly with the new
PostgreSQL notification handling mechanism. The CLI task is only
refreshing the lists of variables to show or not show in Mission
Portal. And that's the same on a superhub, a feeder or any other
hub. So we can now again just rely on the CLI task to be run as
part of an agent run.