Fixed selection of standard_services when used from non-default namespace
Prior to this change, if you were in a non-default namespace and specified
default:standard_service as the service method, the bundle would not be found
because it was looking in the callers namespace and not the default namespace.
This change makes sure that any time standard_services service_method from the
default namespace is requested that the standard_services bundle from the
default namespace is used for the implementation.

Ticket: ENT-5406
Changelog: Title