Removed restriction to re-consider if we have a symlink to an acceptable python
If we only consider the python symlink if there is no file then we break in
cases where a user has changed the python version after the symlink was created.
Now, users will be able to switch python versions and our symlink should follow
or in the worst case fall back to platform-python.