Changed m_inventory dumping behavior to exclude when values is null
Previously an empty values column in m_inventory such as when a
host has not reported would cause corrupt SQL to be dumped.
In this case we should simply not include that host in the
output in a similar way that __inventory is built from v_inventory
in latest schema.

Changelog: Title
Ticket: ENT-5562