Added service states "active" and "inactive" for systemd
This change brings understanding of "active" and "inactive" states for services
when managed by systemd. An "active" service state means the service should be
running. An "inactive" service state means the service should not be running.

NOTE: When "active" or "inactive" service states are used, no promises about a
services state on boot are made.

Ticket: ENT-6074
Changelog: Title