Removed delay in refreshing software installed inventory
This change sets the delay for updating the inventory of currently installed
packages to 0, effectively disabling any delay. This inventory affects
packagesmatching() as well as inventory collected by an Enterprise hub. Without
this change, changes to software made outside CFEngine may not be realized for
up to 60 minutes.

Ticket: ENT-6154
Changelog: Title