Shortened Inventory OS attribute to be more readable
This is the main string used in many places, including
MP UI, pie charts, reports etc.

We want it to be the shortest, most human intuitive and readable way, so:

* Ubuntu 18
* Debian 9
* CentOS 6
* RHEL 7
* SUSE 15
* Windows 2016

We use the old implementation (based on LSB) as a fall back,
but prefer data from os-release, and in some cases we have overrides
for specific platform versions to ensure consistency, even
when the OS maintainers don't agree with each other or even their past
selves.

Even if it isn't perfect everywhere, always, that is okay,
it's supposed to be a shortened pretty string.

Ticket: ENT-6536
Changelog: Title
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>