Suppressed inform output from Enterprise Hub database maintenance operations
The inform output generated by these commands is distracting for new users.
Already the promises were set to be contained in a silent body, this just
extends that silence to suppress from inform. The command results are still
interpreted the same, and if one of these commands is not kept, they will still
emit an error.

Ticket: ENT-6563
Changelog: Title