query service status on FreeBSD using onestatus
If a service is not enabled in rc.conf on FreeBSD, trying to get the status always results in something like:

# /etc/rc.d/ntpd status
Cannot 'status' ntpd. Set ntpd_enable to YES in /etc/rc.conf or use 'onestatus' instead of 'status'.
# echo $?
0

As the exit code is zero, cf-agent would try to stop this service. Using onestatus always returns the wanted exit code, regardless if the service is enabled or not. 

# /etc/rc.d/ntpd onestatus
ntpd is not running.
# echo $?
1