Renamed python symlink to cfengine-selected-python
Package modules that rely on python can be run with either python2 or python3.
We maintain a symlink pointing to a version of python found on the system that
is appropriate. In order to avoid confusion, since CFEngine's bin directory is
commonly found in $PATH we have renamed the symlink so that the symlink is not
used when a user types python on the console.

Ticket: CFE-3512
Changelog: Title
(cherry picked from commit a864d5cced81b37cf0f0ed9e61d392bd6ed96194)