Fixed promise status from package upgrade when architecture specified in promise
This change normalizes reported architectures. Without this patch,
cfengine could not correctly report that the packages installation
promise has been repaired if it is an upgrade (when a previous version
of the package was already installed) or if the architecture has been
explicitly declared in the promise.

Ticket: CFE-3568
Changelog: Title