Remove stale unused copy of u_kept_successful_command body
In <https://gitlab.com/vsa-cfengine/policy-channels/-/issues/4>,
it came up that we have an extra, out-of-date copy of
`body classes u_kept_successful_command body` in lib/common.cf

Remove it so that it doesn't confuse us.

See aforelinked issue for discussion of this and @nickanderson stated it
was probably some artifact left in promises.cf from some time when
update.cf was included in the main policy.