Removed unused plugins directory
The plugins directory has seemingly gone unused since it was introduced to the
MPF in 2013, this change simply removes this directory from being considered for
sharing and updating.

Ticket: CFE-3618
Changelog: Title

The access promise for the plugins directory is retained for now, as during the
upgrade period, there will be old versions of policy that try to copy from this
location and we do not want them to error, causing a cognitive burden while
determining what is wrong.