Restored defintion for def.dir_plugins to fix access promise
The access promsise for the plugin directory should remain since old clients,
e.g. running 3.12.6 policy for example, will still have a promise to copy from
plugins. In order to avoid promise not kept because of no ability to stat the
file, we need to keep the access promise working. But dropping this variable
resulted in skipped access promise.