Switched Travis to using cf-remote for getting master packages
Prior to this change, Travis was fetching packages from builds.cfengine.com
which was previously co-located with cfengine.com. However, this broke when we
deployed the new cfengine.com website into different infrastructure. Instead of
restoring functionality to builds.cfengine.com, I thought it better to use our
shiny new tooling and switched to cf-remote. The Travis build host is Ubuntu
16.04, so, in order to facilitate the change I had to install python 3.6 from
the deadsnakes personal package archive as it is required for both pip3 and cf-remote.

Ticket: ENT-6992
Changelog: None