Changed default behavior of policy update to keep inputs in sync with masterfiles
Prior to this change, the default behavior of the MPF was to only ensure that
files in masterfiles were up to date with the files in inputs. Files in inputs
that did not exist in masterfiles were left undisturbed. To enable sync
behavior (a common user expectation) you had to explicitly define
'cfengine_internal_purge_policies'. Now, if you wish to return to the previous
default behavior, define the class 'cfengine_internal_purge_policies_disabled'.

Ticket: CFE-3662
Changelog: Commit