apt_get package module now uses apt-get for file based installs instead of dpkg
Prior to this change packages promises using the apt_get package module would
use dpkg to handle the case when the path to a package was the promiser. dpkg
does not resolve dependencies, so the result could be a half installed package,
or a package installed with unmet dependencies leaving it in a nonfunctional
state.

This change switches to using apt-get to install the package from a local file,
with apt-get, dependencies are automatically resolved and would be installed
from repositories if available.

Ticket: CFE-3724
Changelog: Title