Added fallback logic for determining installed software version on Windows
Windows software is like the wild west. This change tries to fall back to
InstalledVersion if DisplayVersion is not defined in the registry, if that
fails, it falls back trying to look for keys named the same as the software
itself which while not correct, was seen in the wild for at least one piece of
software.

Ticket: ENT-7501
Changelog: Title