Added class to prevent hub from seeding binary packages for use in self upgrade
Sometimes (especially in testing environments) it's desirable to avoid
downloading packages to master_software_updates on the hub. This can already be
avoided by carefully defining the trigger_upgrade class so that it is not
defined on the hub, but it's helpful to have an explicit way of disabling this
functionality.

Ticket: ENT-7544
Changelog: Title