Reduced scope of report informing of missing systemd service
This change restores 71cb6946b2e45c3466f45c0be29cd377508b22ec from CFE-290 which
limits the scope of the report to cases when the desire is to have the service
functioning. When the desire is to have a service not function, a missing
service fufills this desire. Only in the case of desiring a functional service
does is a missing service problematic.

Ticket: ENT-7360
Changelog: Title