Stopped enforcing permissions of modules in inputs
This change removes explicit enforcement of permissions for modules in inputs.
Instead of explicitly enforcing permissions in inputs, we rely on the default
permissions (600). The previous explicit permissions (755) are un-necessary as
modules are not executed from within the inputs directory and have resulted in
permission flip-flopping in some environments. Permissions on modules in the
modules dir (sys.workdir)/modules are still enforced.

Ticket: ENT-7733
Changelog: Commit