Made apache restart more robust
Changing the apache configuration file can result in apache being unable to
stop/restart when the new config file is in place. Case in point, moving the
httpd.pid. If the config file has been updated, then apachectl will fail to find
a pid in the new loction, being unable to stop or restart.

This change carefully stops apache with the old config when there is a valid
config change and then starts it after the config has been moved into place.

Ticket: ENT-8045
Changelog: Title