Skip invalid feeder dump files during import
Feeder dump files have in their name the hostkey of the feeder.
If this hostkey is not present in the __hubs table then skip
that dump file and log a warning recommending either to delete
the file or add the feeder to superhub.

Ticket: ENT-8229
Changelog: title
(cherry picked from commit ad83552cec64b507251422b692d49f669d56a069)