Remove version condition on FR clean_when_off bundle
It is no longer needed now that we check for the existence
of the FR tables to truncate.
cf_version_minimum is not even support until 3.16 so can't
really be used to check for 3.15 and newer.

Ticket: ENT-8294
Changelog: title
(cherry picked from commit 8597d8b44290d0e301eb94627403bf3fcb2bd43b)