Cleaned up policy related to versions prior to 3.12
Now that 3.12.x is not under standard support, we can remove anything that was
not used in 3.12. Things used in 3.12 need to remain to cover the case where
someone is upgrading to 3.15.x because 3.12.x was supported when 3.15.0 was
released.

This change redacts policy related to cf-consumer and cf-redis-server.

Ticket: CFE-3920
Changelog: Title