Removed policy deprecated by sys.os_release
The variable sys.os_release was introduced in CFEngine 3.11.0. It brought
classification including classes and variables based on the content of
/etc/os-release.

This change simply removes the definition of variables and classes containing
the same information.

- inventory_linux.os_release_info is superceeded by sys.os_release.
- inventory_linux.os_release_id is superseded by sys.os_release[ID].
- inventory_linux.os_release_version is superceeded by sys.os_release[VERSION].

The class default:has_os_release remains as it is used by other MPF policy that
is not directly related to sys.os-release.

Ticket: CFE-3933
Changelog: Title