Adjusted rxdirs to have correct explicit default
Beginning with CFEngine 3.20 rxdirs default has changed from true to false.

This change explicitly sets the value to align with binary defaults. When
running binary versions prior to 3.20 rxdirs is enabled. For versions 3.20 and
greater rxdirs is disabled by default.

Starting with CFEngine 3.18.2, if a perms body sets mode without
specifying rxdirs a warning will be emitted. This change prevents those warnings
for perms bodies that are part of the MPF.

Ticket: CFE-951
Changelog: Title