Use simple numeric comparison in apt_get package module
Previously we used distutils.LooseVersion but distutils is being
deprecated in newer Python versions, such as that available in Ubuntu 22,
so we must migrate to something else.

After looking around a bit and considering our use of LooseVersion it
seems a simple version of this check is all that is needed.

Assuming that apt-get -v always returns simple numeric versions
we compare versions as a list of integers.

Ticket: CFE-3976
Changelog: title