Added policy to patch apachectl for more robust stopping on Enterprise Hubs
The stock apachectl does not wait to see if a signal had the desired effect. In
the case of stopping and restarting this can result in httpd processes being
left running preventing new configuration from taking effect.

This change renders a patched apachectl which is instrumented to briefly wait
for httpd to be fully terminated and takes more aggressive actions to ensure
httpd is fully stopped before exiting.

This script is part of the package but needs to also be handled by the MPF to
cover the case of upgrade where policy from the newer version is running before
binaries from the newer version are running.

Ticket: ENT-8823
Changelog: Title
(cherry picked from commit 70b8210d57e762585150bc62d72958ea2226550a)