Made cmdb update ignore locks
This change is intended to improve the smoothness with respect to unscheduled
agent runs in relation to CMDB data update. Specifically so that the default 1
minute locking period does not prevent an unscheduled agent run from updating
it's CMDB data.

For example, when Trigger agent run button is clicked, a request is made for the
host to do an agent run. If there was an agent less than one minute ago the CMDB
data would not update because of the lock.

Ticket: ENT-8847
Changelog: Title