Fixed directory in which windows agents source packages for upgrade
In 3.18.0 sys.flavor on windows changed so that it includes the major OS version
as is the case for most other platforms. That change broke the expectation of
where to find software during self upgrade. This change simply re-aligns the
source directory used. Since we only build one package for each architecture we
keep using a single directory instead of a separate directory per windows version.

Ticket: ENT-9010
Changelog: Title