Fixed MP Build initial deploy agent run
MP Build deploy of an initial project initiates
an agent run vi the runagent socket listened to by cf-execd.

We need to run the agent twice in this case:
first to set the default:cfengine_internal_masterfiles_update class
second to run masterfiles-stage.sh

In order to make that second run able to activate bundle
cfe_internal_dc_workflow, it must be skipped and not start
a lock in the first agent run.

Ticket: ENT-9625
Changelog: none