Fixed body perms system_owned to account for Windows
This change accounts for windows in the system_owned perms body.

Note: There is a limitation not being able to set file ownership to users other
than the executing user. Since users report said errors when this targets
Administrator or SYSTEM, the owner for Windows is explicitly excluded.

Ticket: ENT-9778
Changelog: Title