Redacted old, useless comment
I don't think that the fact that this was tested on 3.5.x at some point is
really useful. In fact, at this point reading a random snippet it sends me down
a thought path that the policy is quite old, perhaps from the 3.6 era. So, I
think it's best to simply redact this.