Added guards against using regline() in cases where a file may not exist
regline() may throw an error in the event that the probed file does not exist,
as observed in CI in 3.18.x error: In function "regline", error reading from file. (getline: No such file or directory)

To avoid those errors this change guards against use of regline() in case of a
nonexistent at a high level. This instrumentation was not made deep inside
edit_line bundles as it may be useful to surface the potential error in some cases.

Ticket: ENT-9933
Changelog: Title