Fixed suse package_inventory defaults
Before, only redhat or redhat_derived systems were having package_inventory defined.
This caused no software inventory to be reported to an enterprise hub for suse systems.

Previously a check was added for the yum package module written in python that the python version must be 3.x or greater than 2.4.

In this change we have refactored to check the python version for all linux cases of package_inventory.

This check seems sufficient for apt_get, yum and zypper, the three package modules written in python.

Ticket: ENT-10248
Changelog: title