Adjusted dump.sh for multiple runs in between superhub imports
If the dump procedure happens to run either via cf-execd or via manual cf-agent -KI runs
more often than the superhub attempts to import the ln command could fail.

In this case, adding -f (force) makes sense as the end result is the same and what we want: the feeder dump is linked to the proper place for each superhub.

Ticket: ENT-10274
Changelog: title