Fixed cfe_autorun_inventory_aws_ec2_metadata_cache
Before, the cache file was not created because the variable in the regcmp was
not referenceable. default:cfe_autorun_inventory_aws_ec2_metadata_cache.response
is a data container. It's content key is a string, as such, it's not possible to
dereference $(response[content][version]) directly. $(response[content]) must be
extracted into a data container which can be referenced.

Changelog: Fixed cfe_autorun_inventory_aws_ec2_metadata_cache file creation
Ticket: CFE-4221

Signed-off-by: Andreas Gerler <baron@bundesbrandschatzamt.de>